refactor(stella-harbor): split binary location into locate.py
__init__.py is grandfathered at 2199 lines and closed to growth; the
broken-pin refusal pushed it over. locate_binary, _cached_binary and the
binary name move to a package-import-free sibling, re-exported under the
old private name so every call site and test keeps working.

Refs #2098
